Fix/report adjustment

+++ src/app/module/main/main-valuation/models-details/fcfe-details/fcfe-details.component.ts
@@ -144,7 +144,7 @@ loadOnChangeValue(){
         const dialogRef = this.dialog.open(GenericModalBoxComponent,data);
         dialogRef.afterClosed().subscribe((result)=>{
           if (result) {
-            this.fcfeForm.controls['expMarketReturn'].patchValue(parseInt(result?.analystConsensusEstimates))
+            this.fcfeForm.controls['expMarketReturn'].patchValue(parseFloat(result?.analystConsensusEstimates))
             this.snackBar.open('Analyst Estimation Added','OK',{
               horizontalPosition: 'right',
               verticalPosition: 'top',
